--- src/routes/timeattack-goal-routes.js
@@ -1,12 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const TeamGoalTimeAttackController = require('../controllers/timeattack-goal-controller.js');
-
-//const auth = require('../middlewares/auth');
-
-
-router.get('/team/:goal_id/timeattack' , TeamGoalTimeAttackController.getTeamGoalTimeAttack);
-
-
-module.exports = router;
