+++ src/services/timeattack-validation-service.js
@@ -1,7 +1,7 @@
-// 팀미션(타임어택) 진행상황
-const goalModel = require('../models/timeattack-goal-model');
+//팀 목표 인증내역
+const goalModel = require('../models/timeattack-validation-model');
 
-// 목표 ID와 사용자 ID로 목표와 진행 상황을 조회
+// 목표 ID와 사용자 ID로 인증내역 조회
 const getTeamGoalTimeAttack = async (goalId, userId) => {
   if (!goalId || !userId) {
     throw new Error('Goal ID and User ID are required');
